+++ playground/springboot-thymeleaf-full-example/src/main/java/com/kinde/KindeSpringOauthApplication.java
@@ -1,12 +1,36 @@
 package com.kinde;
 
+import io.github.cdimascio.dotenv.Dotenv;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.MapPropertySource;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @SpringBootApplication
 public class KindeSpringOauthApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(KindeSpringOauthApplication.class, args);
+        SpringApplication application = new SpringApplication(KindeSpringOauthApplication.class);
+        application.addInitializers(new DotenvInitializer());
+
+        application.run(args);
+    }
+
+    static class DotenvInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+        @Override
+        public void initialize(ConfigurableApplicationContext applicationContext) {
+            ConfigurableEnvironment environment = applicationContext.getEnvironment();
+            Dotenv dotenv = Dotenv.load();
+
+            Map<String, Object> envMap = new HashMap<>();
+            dotenv.entries().forEach(entry -> envMap.put(entry.getKey(), entry.getValue()));
+
+            environment.getPropertySources().addFirst(new MapPropertySource("dotenvProperties", envMap));
+        }
     }
 }
