+++ playground/springboot-thymeleaf-full-example/src/main/java/com/kinde/oauth/config/CustomOidcUserService.java
@@ -1,5 +1,6 @@
 package com.kinde.oauth.config;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserRequest;
@@ -7,6 +8,8 @@ import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserService;
 import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
 
 import java.util.Collection;
 import java.util.List;
@@ -18,6 +21,13 @@ import java.util.stream.Collectors;
  */
 public class CustomOidcUserService extends OidcUserService {
 
+
+    private final JwtDecoder jwtDecoder;
+
+    public CustomOidcUserService(@Value("${jwk-set-uri}") String issuerUri) {
+        this.jwtDecoder = NimbusJwtDecoder.withJwkSetUri(issuerUri).build();
+    }
+
     /**
      * Loads the OIDC user based on the provided OidcUserRequest, extracting
      * additional authorities from the JWT access token.
@@ -28,24 +38,10 @@ public class CustomOidcUserService extends OidcUserService {
     @Override
     public OidcUser loadUser(OidcUserRequest userRequest) {
         OidcUser oidcUser = super.loadUser(userRequest);
-        String accessToken = userRequest.getAccessToken().getTokenValue();
-        Jwt jwt = parseJwtToken(accessToken);
+        Jwt jwt = jwtDecoder.decode(userRequest.getAccessToken().getTokenValue());
         Collection<GrantedAuthority> authorities = extractAuthoritiesFromJwt(jwt);
 
         return new DefaultOidcUser(authorities, oidcUser.getIdToken(), oidcUser.getUserInfo());
-    }
-
-    /**
-     * Parses the JWT token from the provided token string.
-     *
-     * @param token the JWT token string.
-     * @return the parsed Jwt object.
-     */
-    private Jwt parseJwtToken(String token) {
-        return Jwt.withTokenValue(token)
-                .header("alg", "none")
-                .claim("permissions", List.of("read", "admin"))
-                .build();
     }
 
     /**
@@ -55,7 +51,7 @@ public class CustomOidcUserService extends OidcUserService {
      * @param jwt the parsed Jwt object.
      * @return a collection of GrantedAuthority extracted from the JWT token.
      */
-    private Collection<GrantedAuthority> extractAuthoritiesFromJwt(Jwt jwt) {
+    Collection<GrantedAuthority> extractAuthoritiesFromJwt(Jwt jwt) {
         List<String> permissions = jwt.getClaimAsStringList("permissions");
 
         // Convert permissions to Spring Security roles
