+++ playground/springboot-thymeleaf-full-example/src/main/java/com/kinde/oauth/config/WebClientConfiguration.java
@@ -8,25 +8,35 @@ import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager;
 import org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction;
 import org.springframework.web.reactive.function.client.WebClient;
 
+/**
+ * Configuration class for setting up a {@link WebClient} to interact with
+ * OAuth2-protected resources.
+ */
 @Configuration
 public class WebClientConfiguration {
 
-	@Value("${user-profile-uri}")
-	String userprofileUri;
-	@Value("${logout-uri}")
-	String logoutUri;
+    @Value("${user-profile-uri}")
+    String userprofileUri;
 
-	@Bean
-	@Qualifier(value = "userProfile")
-	WebClient userProfileClient(OAuth2AuthorizedClientManager authorizedClientManager) {
+    /**
+     * Creates a {@link WebClient} bean specifically for accessing the user profile
+     * resource, configured with OAuth2 authorization support.
+     *
+     * @param authorizedClientManager the {@link OAuth2AuthorizedClientManager} used to
+     *                                manage OAuth2 authorized clients.
+     * @return a configured {@link WebClient} instance.
+     */
+    @Bean
+    @Qualifier(value = "userProfile")
+    WebClient userProfileClient(OAuth2AuthorizedClientManager authorizedClientManager) {
 
-		ServletOAuth2AuthorizedClientExchangeFilterFunction oauth2 = new ServletOAuth2AuthorizedClientExchangeFilterFunction(
-				authorizedClientManager);
-		oauth2.setDefaultOAuth2AuthorizedClient(true);
+        ServletOAuth2AuthorizedClientExchangeFilterFunction oauth2 = new ServletOAuth2AuthorizedClientExchangeFilterFunction(
+                authorizedClientManager);
+        oauth2.setDefaultOAuth2AuthorizedClient(true);
 
-		return WebClient.builder()
-				.baseUrl(this.userprofileUri)
-				.apply(oauth2.oauth2Configuration())
-				.build();
-	}
+        return WebClient.builder()
+                .baseUrl(this.userprofileUri)
+                .apply(oauth2.oauth2Configuration())
+                .build();
+    }
 }
