+++ playground/springboot-thymeleaf-full-example/src/main/java/com/kinde/oauth/service/KindeService.java
@@ -1,5 +1,7 @@
 package com.kinde.oauth.service;
 
+import com.kinde.KindeClient;
+import com.kinde.KindeClientBuilder;
 import com.kinde.oauth.model.KindeProfile;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -13,21 +15,37 @@ import java.util.HashSet;
 
 import static org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction.oauth2AuthorizedClient;
 
+/**
+ * Service class to handle interactions with the Kinde OAuth2 authorization service
+ * and retrieve user information.
+ */
 @Service
 public class KindeService {
 
     private final OAuth2AuthorizedClientService authorizedClientService;
     private final WebClient userProfileClient;
 
+    /**
+     * Constructor to initialize the KindeService.
+     *
+     * @param authorizedClientService the service for managing authorized OAuth2 clients.
+     * @param userProfileClient       the {@link WebClient} used to fetch the user profile.
+     */
     public KindeService(OAuth2AuthorizedClientService authorizedClientService,
                         WebClient userProfileClient) {
         this.authorizedClientService = authorizedClientService;
         this.userProfileClient = userProfileClient;
     }
 
-    public KindeProfile loadDashboard(Authentication authentication) {
+    /**
+     * Loads the user dashboard by fetching the OAuth2 authorized client and retrieving
+     * the user profile data.
+     *
+     * @return a {@link KindeProfile} object containing user details.
+     */
+    public KindeProfile loadDashboard() {
 
-        OAuth2AuthorizedClient authorizedClient = getoAuth2AuthorizedClient(authentication);
+        OAuth2AuthorizedClient authorizedClient = getOAuth2AuthorizedClient(SecurityContextHolder.getContext().getAuthentication());
 
         // Extracting user profile via Spring Webflux client
         String userprofile = this.userProfileClient
@@ -47,7 +65,13 @@ public class KindeService {
                 .build();
     }
 
-    private OAuth2AuthorizedClient getoAuth2AuthorizedClient(Authentication authentication) {
+    /**
+     * Retrieves the OAuth2 authorized client for the authenticated user.
+     *
+     * @param authentication the current user's {@link Authentication} object.
+     * @return the {@link OAuth2AuthorizedClient} for the user.
+     */
+    private OAuth2AuthorizedClient getOAuth2AuthorizedClient(Authentication authentication) {
         return authorizedClientService.loadAuthorizedClient("kinde", authentication.getName());
     }
 }
