+++ playground/springboot-thymeleaf-full-example/src/main/resources/application.properties
@@ -2,30 +2,32 @@ spring.application.name=kinde-spring-oauth
 
 server.port=8081
 
+spring.config.import=optional:file:.env
+
 server.error.whitelabel.enabled=false
 server.error.path=/error
 
 logging.level.org.springframework.security=trace
 
-spring.security.oauth2.client.registration.kinde.client-id=< replace >
-spring.security.oauth2.client.registration.kinde.client-secret=< replace >
-spring.security.oauth2.client.registration.kinde.scope=openid,profile,email
-spring.security.oauth2.client.registration.kinde.redirect-uri=http://localhost:8081/login/oauth2/code/kinde-provider
-spring.security.oauth2.client.registration.kinde.authorization-grant-type=authorization_code
+spring.security.oauth2.client.registration.kinde.client-id=${KINDE_CLIENT_ID}
+spring.security.oauth2.client.registration.kinde.client-secret=${KINDE_CLIENT_SECRET}
+spring.security.oauth2.client.registration.kinde.scope=${KINDE_SCOPES}
+spring.security.oauth2.client.registration.kinde.redirect-uri=${KINDE_REDIRECT_URI}
+spring.security.oauth2.client.registration.kinde.authorization-grant-type=${KINDE_GRANT_TYPE}
 spring.security.oauth2.client.registration.kinde.client-name=Kinde
 spring.security.oauth2.client.registration.kinde.provider=kinde
 
-spring.security.oauth2.client.provider.kinde.authorization-uri=https://< replace >.kinde.com/oauth2/auth
-spring.security.oauth2.client.provider.kinde.token-uri=https://< replace >.kinde.com/oauth2/token
-spring.security.oauth2.client.provider.kinde.user-info-uri=https://< replace >.kinde.com/oauth2/v2/user_profile
-spring.security.oauth2.client.provider.kinde.jwk-set-uri=https://< replace >.kinde.com/.well-known/jwks.json
+spring.security.oauth2.client.provider.kinde.authorization-uri=https://koman.kinde.com/oauth2/auth
+spring.security.oauth2.client.provider.kinde.token-uri=https://koman.kinde.com/oauth2/token
+spring.security.oauth2.client.provider.kinde.user-info-uri=https://koman.kinde.com/oauth2/v2/user_profile
+spring.security.oauth2.client.provider.kinde.jwk-set-uri=https://koman.kinde.com/.well-known/jwks
 spring.security.oauth2.client.provider.kinde.user-name-attribute=sub
 
-spring.security.oauth2.resourceserver.jwt.issuer-uri=https://< replace >.kinde.com
-spring.security.oauth2.resourceserver.jwt.jwk-set-uri=https://< replace >.kinde.com/.well-known/jwks.json
+spring.security.oauth2.resourceserver.jwt.issuer-uri=${KINDE_DOMAIN}
 
-user-profile-uri=https://< replace >.kinde.com/oauth2/user_profile
-logout-uri=https://< replace >.kinde.com/logout
+jwk-set-uri=https://koman.kinde.com/.well-known/jwks
+user-profile-uri=https://koman.kinde.com/oauth2/user_profile
+logout-uri=https://koman.kinde.com/logout
 
 spring.thymeleaf.cache=false
 spring.thymeleaf.prefix=classpath:/templates/
