+++ playground/springboot-thymeleaf-full-example/src/main/resources/templates/403.html
@@ -18,10 +18,24 @@
             <p class="text-display-1 hero-title">
                 Nope! 403 - Access denied
             </p>
-
-            <a href="/dashboard" rel="noreferrer" class="btn btn-light btn-big">
-                Home
-            </a>
+            <p class="text-body-1 hero-tagline">
+                You attempted to access a page you do not have authorization for.
+            </p>
+            <p class="text-body-2">
+                Please check your Kinde profile and add the necessary roles and permissions.
+            </p>
+            <p>
+                <br>
+                <br>
+                <a href="https://docs.kinde.com/manage-users/roles-and-permissions/user-roles/?r=search#_top"
+                   target="_blank" rel="noreferrer"
+                   class="btn btn-light btn-small">
+                    Roles & Permissions
+                </a>
+                <a href="/dashboard" rel="noreferrer" class="btn btn-light btn-small">
+                    Home
+                </a>
+            </p>
         </div>
     </div>
 </main>
@@ -30,7 +44,7 @@
     <div class="container">
         <p class="footer-tagline text-body-3">
             Browse Kinde
-            <a class="link" href="https://kinde.com/docs">docs</a>
+            <a class="link" target="_blank" href="https://kinde.com/docs">docs</a>
         </p>
 
         <small class="text-subtle">© 2024 KindeAuth, Inc. All rights reserved</small>
