+++ playground/springboot-thymeleaf-full-example/src/main/resources/templates/dashboard.html
@@ -123,7 +123,7 @@
             <strong class="text-heading-2">KindeAuth</strong>
             <p class="footer-tagline text-body-3">
                 Browse Kinde
-                <a class="link" href="https://kinde.com/docs">docs</a>
+                <a class="link" target="_blank" href="https://kinde.com/docs">docs</a>
             </p>
 
             <small class="text-subtle">© 2024 KindeAuth, Inc. All rights reserved</small>
