+++ playground/springboot-thymeleaf-full-example/src/main/resources/templates/home.html
@@ -37,7 +37,7 @@
     <div class="container">
         <p class="footer-tagline text-body-3">
             Browse Kinde
-            <a class="link" href="https://kinde.com/docs">docs</a>
+            <a class="link" target="_blank" href="https://kinde.com/docs">docs</a>
         </p>
 
         <small class="text-subtle">© 2024 KindeAuth, Inc. All rights reserved</small>
