--- playground/springboot-thymeleaf-full-example/src/src/main/java/com/kinde/oauth/model/KindeProfile.java
@@ -1,17 +1,0 @@
-package com.kinde.oauth.model;
-
-import lombok.Builder;
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-
-import java.util.Set;
-
-@Data
-@Builder
-public class KindeProfile {
-    private String idToken;
-    private String accessToken;
-    private String fullName;
-    private Set<GrantedAuthority> roles;
-    private String userProfile;
-}
