--- playground/springboot-thymeleaf-full-example/src/src/main/resources/templates/403.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="en">
-<head>
-    <title>KindeAuth</title>
-    <link rel="stylesheet" type="text/css" th:href="@{/css/index.css}"/>
-
-</head>
-<body>
-<header>
-    <nav class="nav container">
-        <h1 class="text-display-3">KindeAuth with Spring Boot</h1>
-    </nav>
-</header>
-
-<main>
-    <div class="container">
-        <div class="card hero">
-            <p class="text-display-1 hero-title">
-                Nope! 403 - Access denied
-            </p>
-            <p class="text-body-1 hero-tagline">
-                You attempted to access a page you do not have authorization for.
-            </p>
-            <p class="text-body-2">
-                Please check your Kinde profile and add the necessary roles and permissions.
-            </p>
-            <p>
-                <br>
-                <br>
-                <a href="https://docs.kinde.com/manage-users/roles-and-permissions/user-roles/?r=search#_top"
-                   target="_blank" rel="noreferrer"
-                   class="btn btn-light btn-small">
-                    Roles & Permissions
-                </a>
-                <a href="/dashboard" rel="noreferrer" class="btn btn-light btn-small">
-                    Home
-                </a>
-            </p>
-        </div>
-    </div>
-</main>
-
-<footer class="footer">
-    <div class="container">
-        <p class="footer-tagline text-body-3">
-            Browse Kinde
-            <a class="link" target="_blank" href="https://kinde.com/docs">docs</a>
-        </p>
-
-        <small class="text-subtle">© 2024 KindeAuth, Inc. All rights reserved</small>
-    </div>
-</footer>
-</body>
-</html>
