--- playground/springboot-thymeleaf-full-example/src/src/main/resources/templates/error.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="en">
-<head>
-    <title>KindeAuth</title>
-    <link rel="stylesheet" type="text/css" th:href="@{/css/index.css}"/>
-
-</head>
-<body>
-<header>
-    <nav class="nav container">
-        <h1 class="text-display-3">KindeAuth with Spring Boot</h1>
-    </nav>
-</header>
-
-<main>
-    <div class="container">
-        <div class="card hero">
-            <p class="text-display-1 hero-title">
-                ERROR
-            </p>
-
-            <a href="/dashboard" rel="noreferrer" class="btn btn-light btn-big">
-                Home
-            </a>
-        </div>
-    </div>
-</main>
-
-<footer class="footer">
-    <div class="container">
-        <p class="footer-tagline text-body-3">
-            Browse Kinde
-            <a class="link" target="_blank" href="https://kinde.com/docs">docs</a>
-        </p>
-
-        <small class="text-subtle">© 2024 KindeAuth, Inc. All rights reserved</small>
-    </div>
-</footer>
-</body>
-</html>
